Unify behavior of MultiCommand
The outer parser was showing different help messages compared to the
inner ones. For example, it showed "Syntax:" when the inner parsers
showed "usage:". Also, on the main level "--help" was never supported,
but is now. When usage is printed as a result of an error, sys.stderr is
used, but when usage is printed on user request (i.e., "--help"),
sys.stdout is used -- just like the argparse class does it, too.